pyproject.toml
[project]
name = "smart-artifact-parser"
version = "0.1.0"
description = "Extract structured medical information from documents using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.15.0",
    "anthropic>=0.40.0",
    "pydantic>=2.10.0",
    "typer>=0.15.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
smart-parser = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
